Unit tests expanded by exceptions handling. Covarage 96% (br.), 78% (mut.). Message getBody() not tested yet.
