+++ src/test/java/com/acme/edu/ooad/controller/LoggerControllerTest.java
@@ -14,11 +14,48 @@ public class LoggerControllerTest {
     LoggerController controllerSut = new LoggerController(saverMock);
 
     @Test
-    public void shouldSaveLastLoggedMessageWhenFlushAndLastLoggedMessageIsNotNull() throws FlushException, SaveException {
-        controllerSut.lastLoggedMessage = mock(Message.class);
-        controllerSut.flush();
+    public void shouldSaveLastLoggedMessageWhenFlushAndLastLoggedMessageIsNotNull(){
+        Message messageStub = mock(Message.class);
+        controllerSut.lastLoggedMessage = messageStub;
+        try {
+            controllerSut.flush();
+        } catch (FlushException e) {
+            e.printStackTrace();
+        }
 
-        verify(saverMock, times(1)).save(any(Message.class));
+        try {
+            verify(saverMock, times(1)).save(messageStub);
+        } catch (SaveException e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Test
+    public void shouldGetFlushErrorWhenSaverGetError() throws SaveException{
+        Message emptyMessageStub = mock(Message.class);
+        controllerSut.lastLoggedMessage = emptyMessageStub;
+        doThrow(SaveException.class).when(saverMock).save(emptyMessageStub);
+
+        assertThrows(
+                FlushException.class,
+                () -> controllerSut.flush()
+        );
+    }
+
+    @Test
+    public void shouldGetLogErrorWhenSaverGetError() throws SaveException{
+        Message messageStub = mock(Message.class);
+
+        Message lastLoggedMessageStub = mock(Message.class);
+        when(lastLoggedMessageStub.getInstanceToPrint(any())).thenReturn(messageStub);
+        controllerSut.lastLoggedMessage = lastLoggedMessageStub;
+
+        doThrow(SaveException.class).when(saverMock).save(messageStub);
+
+        assertThrows(
+                LogException.class,
+                () -> controllerSut.log(messageStub)
+        );
     }
 
 //    @Test
@@ -41,14 +78,18 @@ public class LoggerControllerTest {
 //    }
 
     @Test
-    public void shouldSaveInstanceToPrintWhenLogAndLastLoggedMessageIsNotNull() throws LogException {
+    public void shouldSaveInstanceToPrintWhenLogAndLastLoggedMessageIsNotNull(){
         Message logMessageStub = mock(Message.class);
         controllerSut.lastLoggedMessage = mock(Message.class);
         Message saveMessageStub = mock(Message.class);
         when(controllerSut.lastLoggedMessage.getInstanceToPrint(logMessageStub))
                 .thenReturn(saveMessageStub);
 
-        controllerSut.log(logMessageStub);
+        try {
+            controllerSut.log(logMessageStub);
+        } catch (LogException e) {
+            e.printStackTrace();
+        }
         try {
             verify(saverMock, times(1)).save(saveMessageStub);
         } catch (SaveException e) {
@@ -57,14 +98,18 @@ public class LoggerControllerTest {
     }
 
     @Test
-    public void shouldUpdateLastLoggedMessageWhenLogAndLastLoggedMessageIsNotNull() throws LogException {
+    public void shouldUpdateLastLoggedMessageWhenLogAndLastLoggedMessageIsNotNull(){
         Message logMessageStub = mock(Message.class);
         Message newInstanceMessageStub = mock(Message.class);
         controllerSut.lastLoggedMessage = mock(Message.class);
         when(controllerSut.lastLoggedMessage.getNewInstance(logMessageStub))
                 .thenReturn(newInstanceMessageStub);
 
-        controllerSut.log(logMessageStub);
+        try {
+            controllerSut.log(logMessageStub);
+        } catch (LogException e) {
+            e.printStackTrace();
+        }
         assertEquals(newInstanceMessageStub, controllerSut.lastLoggedMessage);
     }
 }
