+++ src/test/java/com/acme/edu/ooad/saver/SaverTest.java
@@ -3,15 +3,16 @@ package com.acme.edu.ooad.saver;
 import com.acme.edu.ooad.exception.SaveException;
 import com.acme.edu.ooad.message.Message;
 import org.junit.jupiter.api.Test;
+import com.acme.edu.SysoutCaptureAndAssertionAbility;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class SaverTest {
+public class SaverTest implements SysoutCaptureAndAssertionAbility{
+    Saver saver = new ConsoleSaver();
     @Test
     public void shouldGetErrorWhenMessageIsNull() {
-        Saver saver = new ConsoleSaver();
         Message nullMessage = null;
 
         assertThrows(
@@ -22,12 +23,28 @@ public class SaverTest {
 
     @Test
     public void shouldGetErrorWhenMessageIsEmpty() {
-        Saver saver = new ConsoleSaver();
-        Message emptyMessage = mock(Message.class);
-        when(emptyMessage.getBody()).thenReturn("");
+        Message emptyMessageStub = mock(Message.class);
+        when(emptyMessageStub.getBody()).thenReturn("");
+
         assertThrows(
                 SaveException.class,
-                () -> saver.save(emptyMessage)
+                () -> saver.save(emptyMessageStub)
         );
+    }
+
+    @Test
+    public void shouldPrintWhenSaveMessage() {
+        resetOut();
+        captureSysout();
+
+        Message messageStub = mock(Message.class);
+        when(messageStub.getBody()).thenReturn("message to print");
+        when(messageStub.toString()).thenReturn("message to print");
+        try {
+            saver.save(messageStub);
+        } catch (SaveException e) {
+            e.printStackTrace();
+        }
+        assertSysoutEquals("message to print"+System.lineSeparator());
     }
 }
